+++ src/components/SettingsPropsMain.vue
@@ -8,7 +8,8 @@
     const props = defineProps({
         type: String,
         typeIdx: Number,
-        section: String
+        section: String,
+        devices: Object
     })
 
     const subSections = [
@@ -33,9 +34,9 @@
 
     const listItemDefs = {
         level: [
-            { type: 'defs', required: false, prop: 'show', def: 'reading:value:show', help: 'main-element-show' },
-            { type: 'defs', required: false, prop: 'divider', def: 'reading:value:divider', help: 'main-element-divider' },
-            { type: 'defs', required: false, prop: 'height', def: 'reading:value:height', help: 'main-element-height' },
+            { type: 'defs', required: false, prop: 'show', def: 'reading:value:show', help: 'main-element-show', assist: 'props' },
+            { type: 'defs', required: false, prop: 'divider', def: 'reading:value:divider', help: 'main-element-divider', assist: 'props' },
+            { type: 'defs', required: false, prop: 'height', def: 'reading:value:height', help: 'main-element-height', assist: 'props' },
             { type: 'types', required: false, prop: 'left1', help: 'main-element-typ' },
             { type: 'types', required: false, prop: 'left2', help: 'main-element-typ' },
             { type: 'types', required: false, prop: 'mid', help: 'main-element-typ' },
@@ -43,50 +44,50 @@
             { type: 'types', required: false, prop: 'right2', help: 'main-element-typ' }
         ],
         btn: [
-            { type: 'defs', required: false, prop: 'btn', def: 'reading:value:icon:disabled:color:variant', help: 'level-element-button-btn' },
-            { type: 'defs', required: false, prop: 'status', def: 'reading:value:level:color:min:max:reverse', help: 'level-element-button-status' },
-            { type: 'defs', required: false, prop: 'click', def: 'reading:value:cmd:type', help: 'level-element-button-click' },
-            { type: 'defs', required: false, prop: 'longClick', def: 'reading:value:cmd:type', help: 'level-element-button-longClick' },
-            { type: 'defs', required: false, prop: 'longRelease', def: 'reading:value:cmd:type', help: 'level-element-button-longRelease' },
-            { type: 'defs', required: false, prop: 'divider', def: 'reading:value:divider', help: 'level-element-divider' },
-            { type: 'defs', required: false, prop: 'size', def: 'reading:value:size', help: 'level-element-size' }
+            { type: 'defs', required: false, prop: 'btn', def: 'reading:value:icon:disabled:color:variant', help: 'level-element-button-btn', assist: 'props' },
+            { type: 'defs', required: false, prop: 'status', def: 'reading:value:level:color:min:max:reverse', help: 'level-element-button-status', assist: 'props' },
+            { type: 'defs', required: false, prop: 'click', def: 'reading:value:cmd:type', help: 'level-element-button-click', assist: 'props' },
+            { type: 'defs', required: false, prop: 'longClick', def: 'reading:value:cmd:type', help: 'level-element-button-longClick', assist: 'props' },
+            { type: 'defs', required: false, prop: 'longRelease', def: 'reading:value:cmd:type', help: 'level-element-button-longRelease', assist: 'props' },
+            { type: 'defs', required: false, prop: 'divider', def: 'reading:value:divider', help: 'level-element-divider', assist: 'props' },
+            { type: 'defs', required: false, prop: 'size', def: 'reading:value:size', help: 'level-element-size', assist: 'props' }
         ],
         slider: [
-            { type: 'defs', required: false, prop: 'slider', def: 'reading:value:cmd:current:color:min:max:steps:reverse:size:vertical', help: 'level-element-slider-slider' },
-            { type: 'defs', required: false, prop: 'divider', def: 'reading:value:divider', help: 'level-element-divider' },
-            { type: 'defs', required: false, prop: 'size', def: 'reading:value:size', help: 'level-element-size' }
+            { type: 'defs', required: false, prop: 'slider', def: 'reading:value:cmd:current:color:min:max:steps:reverse:size:vertical', help: 'level-element-slider-slider', assist: 'props' },
+            { type: 'defs', required: false, prop: 'divider', def: 'reading:value:divider', help: 'level-element-divider', assist: 'props' },
+            { type: 'defs', required: false, prop: 'size', def: 'reading:value:size', help: 'level-element-size', assist: 'props' }
         ],
         image: [
-            { type: 'defs', required: false, prop: 'image', def: 'reading:value:source:height', help: 'level-element-image-image' },
-            { type: 'defs', required: false, prop: 'divider', def: 'reading:value:divider', help: 'level-element-divider' },
-            { type: 'defs', required: false, prop: 'size', def: 'reading:value:size', help: 'level-element-size' }
+            { type: 'defs', required: false, prop: 'image', def: 'reading:value:source:height', help: 'level-element-image-image', assist: 'props' },
+            { type: 'defs', required: false, prop: 'divider', def: 'reading:value:divider', help: 'level-element-divider', assist: 'props' },
+            { type: 'defs', required: false, prop: 'size', def: 'reading:value:size', help: 'level-element-size', assist: 'props' }
         ],
         menu: [
-            { type: 'defs', required: false, prop: 'btn', def: 'reading:value:icon:disabled:color:variant', help: 'level-element-menu-btn' },    
-            { type: 'defs', required: false, prop: 'menu', def: 'reading:value:name:cmd', help: 'level-element-menu-menu' },
-            { type: 'defs', required: false, prop: 'divider', def: 'reading:value:divider', help: 'level-element-divider' },
-            { type: 'defs', required: false, prop: 'size', def: 'reading:value:size', help: 'level-element-size' }
+            { type: 'defs', required: false, prop: 'btn', def: 'reading:value:icon:disabled:color:variant', help: 'level-element-menu-btn', assist: 'props' },    
+            { type: 'defs', required: false, prop: 'menu', def: 'reading:value:name:cmd', help: 'level-element-menu-menu', assist: 'props' },
+            { type: 'defs', required: false, prop: 'divider', def: 'reading:value:divider', help: 'level-element-divider', assist: 'props' },
+            { type: 'defs', required: false, prop: 'size', def: 'reading:value:size', help: 'level-element-size', assist: 'props' }
         ],
         info: [
-            { type: 'defs', required: false, prop: 'text', def: 'reading:value:text:format', help: 'level-element-info-text' },
-            { type: 'defs', required: false, prop: 'icon', def: 'reading:value:icon:color:size', help: 'level-element-info-icon' },
-            { type: 'defs', required: false, prop: 'status', def: 'reading:value:level:color:min:max:reverse:linear', help: 'level-element-info-status' },
-            { type: 'defs', required: false, prop: 'text2', def: 'reading:value:text:format', help: 'level-element-info-text' },
-            { type: 'defs', required: false, prop: 'text3', def: 'reading:value:text:format', help: 'level-element-info-text' },
-            { type: 'defs', required: false, prop: 'divider', def: 'reading:value:divider', help: 'level-element-divider' },
-            { type: 'defs', required: false, prop: 'size', def: 'reading:value:size', help: 'level-element-size' }
+            { type: 'defs', required: false, prop: 'text', def: 'reading:value:text:format', help: 'level-element-info-text', assist: 'props' },
+            { type: 'defs', required: false, prop: 'icon', def: 'reading:value:icon:color:size', help: 'level-element-info-icon', assist: 'props' },
+            { type: 'defs', required: false, prop: 'status', def: 'reading:value:level:color:min:max:reverse:linear', help: 'level-element-info-status', assist: 'props' },
+            { type: 'defs', required: false, prop: 'text2', def: 'reading:value:text:format', help: 'level-element-info-text', assist: 'props' },
+            { type: 'defs', required: false, prop: 'text3', def: 'reading:value:text:format', help: 'level-element-info-text', assist: 'props' },
+            { type: 'defs', required: false, prop: 'divider', def: 'reading:value:divider', help: 'level-element-divider', assist: 'props' },
+            { type: 'defs', required: false, prop: 'size', def: 'reading:value:size', help: 'level-element-size', assist: 'props' }
         ],
         chart: [
-            { type: 'defs', required: false, prop: 'serie', def: 'reading:value:data:name:digits:suffix:type', help: 'level-element-chart-serie' },
+            { type: 'defs', required: false, prop: 'serie', def: 'reading:value:data:name:digits:suffix:type', help: 'level-element-chart-serie', assist: 'props' },
             { type: 'json', required: false, prop: 'options', def: '', help: 'level-element-chart-options' },
             { type: 'json', required: false, prop: 'options2', def: '', help: 'level-element-chart-options' },
-            { type: 'defs', required: false, prop: 'divider', def: 'reading:value:divider', help: 'level-element-divider' },
-            { type: 'defs', required: false, prop: 'size', def: 'reading:value:size', help: 'level-element-size' }
+            { type: 'defs', required: false, prop: 'divider', def: 'reading:value:divider', help: 'level-element-divider', assist: 'props' },
+            { type: 'defs', required: false, prop: 'size', def: 'reading:value:size', help: 'level-element-size', assist: 'props' }
         ],
         colorpicker: [
-            { type: 'defs', required: false, prop: 'picker', def: 'reading:value:cmd:current', help: 'level-element-colorpicker-picker' },
-            { type: 'defs', required: false, prop: 'divider', def: 'reading:value:divider', help: 'level-element-divider' },
-            { type: 'defs', required: false, prop: 'size', def: 'reading:value:size', help: 'level-element-size' }
+            { type: 'defs', required: false, prop: 'picker', def: 'reading:value:cmd:current', help: 'level-element-colorpicker-picker', assist: 'props' },
+            { type: 'defs', required: false, prop: 'divider', def: 'reading:value:divider', help: 'level-element-divider', assist: 'props' },
+            { type: 'defs', required: false, prop: 'size', def: 'reading:value:size', help: 'level-element-size', assist: 'props' }
         ]
     }
 
@@ -99,7 +100,7 @@
     }
 
     const listType = computed(() => {
-        let res = fhem.getEl(fhem.app.config[props.type][props.typeIdx][props.section][lvl.value - 1], ['level', subSections[subSectionIdx.value],'name'])
+        let res = fhem.getEl(fhem.app.config[props.type][props.typeIdx][props.section][lvl.value - 1], ['level', subSections[subSectionIdx.value]['name']])
 
         return res
     })
@@ -244,7 +245,9 @@
                         :mainSection="subSections[subSectionIdx].name"
                         :prop="def.prop" 
                         :propDef="def.def"
-                        :propHelp="def.help">
+                        :propHelp="def.help"
+                        :propAssist="def.assist"
+                        :propAssistDevices="props.devices">
                     </SettingsPropsMainItem>
                 </v-expansion-panel-text>
             </v-expansion-panel>